requirements.txt
requests>=2.25.0
pandas>=1.2.0
python-dotenv>=0.15.0
readability-metrics>=1.4.5
nltk>=3.6.0
